backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Supabase - All-in-One (Database + Storage + Auth)
supabase==2.3.4
postgrest==0.13.2
psycopg2-binary==2.9.9

# Database ORM
sqlalchemy==2.0.25
alembic==1.13.1

# Vector Database & Embeddings
pgvector==0.2.4
sentence-transformers==2.3.1
openai==1.10.0

# LlamaIndex for RAG
llama-index==0.10.0
llama-index-vector-stores-supabase==0.1.3
llama-index-embeddings-openai==0.1.6
llama-index-llms-groq==0.1.3

# LLM APIs
groq==0.4.2
tiktoken==0.5.2

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1
httpx==0.26.0

# File Processing
PyPDF2==3.0.1
pypdf==4.0.1

# Monitoring & Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==24.1.1
flake8==7.0.0
